+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 
 ### Fixes
 
+- Fix an issue in the AppleScript support where the `print` command is ignored.
 - Fix an issue where the Settings window was not accessible from the Window menu.
 - Fix an issue on macOS 15 where newly created documents are incorrectly positioned when the frontmost document is tiled.
 
@@ -44,7 +45,6 @@
 ### Fixes
 
 - Fix an issue in the Appearance settings pane where the alert about potential non-monospace fonts keeps appearing even after changing only the font size.
-